// libs/shell/shared/src/haqq-proto/cosmos/base/v1beta1/coin_pb.ts
export class Coin extends Message<Coin> {
  /**
   * @generated from field: string denom = 1;
   */
  denom = '';

  /**
   * @generated from field: string amount = 2;
   */
  amount = '';

  constructor(data?: PartialMessage<Coin>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = 'cosmos.base.v1beta1.Coin';
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: 'denom', kind: 'scalar', T: 9 /* ScalarType.STRING */ },
    { no: 2, name: 'amount', kind: 'scalar', T: 9 /* ScalarType.STRING */ },
  ]);

  static fromBinary(
    bytes: Uint8Array,
    options?: Partial<BinaryReadOptions>,
  ): Coin {
    return new Coin().fromBinary(bytes, options);
  }

  static fromJson(
    jsonValue: JsonValue,
    options?: Partial<JsonReadOptions>,
  ): Coin {
    return new Coin().fromJson(jsonValue, options);
  }

  static fromJsonString(
    jsonString: string,
    options?: Partial<JsonReadOptions>,
  ): Coin {
    return new Coin().fromJsonString(jsonString, options);
  }

  static equals(
    a: Coin | PlainMessage<Coin> | undefined,
    b: Coin | PlainMessage<Coin> | undefined,
  ): boolean {
    return proto3.util.equals(Coin, a, b);
  }
}